requirements.txt
Django
django-localflavor
django-widget-tweaks
Black
